Ajout des pages erreurs.

+++ blueprint/cas/gestion_cas.py
@@ -1,4 +1,4 @@
-from flask import Blueprint, render_template, request, redirect, url_for, flash, session
+from flask import Blueprint, render_template, request, redirect, url_for, abort, session
 from bd import db
 from blueprint.cas.models.cas import Cas
 from blueprint.cas.models.cas import Regions
@@ -12,20 +12,21 @@ routes_cas = Blueprint('gestion_cas', __name__,
 @routes_cas.route('/')
 def accueil():
     regions = Regions.query.all()
-    cas = {}
-    for region in regions:
-        cas[region.nom] = Cas.query.filter_by(region_id=region.id).count()
-    return render_template('liste_region.html', cas=cas, )
+    return render_template('liste_region.html', regions=regions, number=format_number)
 
 
-@routes_cas.route('/cas')
+@ routes_cas.route('/cas')
 def liste_admin():
+    if not session.get('admin'):
+        abort(403)
     cas = Cas.query.all()
-    return render_template('liste_admin.html', lesCas=cas, regions=Regions)
+    return render_template('liste_admin.html', lesCas=cas, regions=Regions, date=format_date)
 
 
-@routes_cas.route('/creer', methods=['GET', 'POST'])
+@ routes_cas.route('/creer', methods=['GET', 'POST'])
 def creer_cas():
+    if not session.get('compte_id'):
+        abort(401)
     if request.method == 'POST':
         cas = Cas(
             nom=request.form.get('nom'),
@@ -34,17 +35,19 @@ def creer_cas():
             compte=session.get('compte_id'),
         )
         if len(cas.erreurs) > 0:
-            # flash(erreurs, 'danger')
             return render_template('saisie.html', cas=cas, messages=cas.erreurs, regions=Regions.query.order_by(Regions.nom).all())
         db.session.add(cas)
         db.session.commit()
-        # flash('Cas créé avec succès !')
-        return redirect('/cas/creer', code=302)
-    return render_template('saisie.html', regions=Regions.query.order_by(Regions.nom).all(), connected=session.get('connected'))
+        return redirect(url_for('gestion_cas.creer_cas'), code=302)
+    return render_template('saisie.html', regions=Regions.query.order_by(Regions.nom).all())
 
 
-@routes_cas.route('/<int:id>/delete', methods=['GET', 'POST'])
+@ routes_cas.route('/<int:id>/delete', methods=['GET', 'POST'])
 def supprimer_cas(id):
+    if not session.get('compte_id'):
+        abort(401)
+    if not session.get('admin'):
+        abort(403)
     if request.method == 'POST':
         id = 5
         Cas.query.filter_by(id=id).delete()
